Configurações de Tela Dicas
- Modificadas algumas regras de Interface do Studio
- Adicionado exibirDicasInterface as configurações de usuário
