+++ nbproject/project.properties
@@ -101,7 +101,7 @@ javadoc.splitindex=true
 javadoc.use=true
 javadoc.version=false
 javadoc.windowtitle=
-main.class=br.univali.ps.ui.TelaPrincipal
+main.class=br.univali.ps.ui.telas.TelaPrincipal
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
