+++ src/br/univali/ps/ui/telas/TelaOpcoesExecucao.java
@@ -1,0 +1,242 @@
+package br.univali.ps.ui.telas;
+
+import br.univali.ps.nucleo.Configuracoes;
+import br.univali.portugol.nucleo.Programa;
+import br.univali.ps.ui.util.IconFactory;
+import java.io.IOException;
+import java.util.Collections;
+import java.util.List;
+import javax.imageio.ImageIO;
+import javax.swing.DefaultComboBoxModel;
+import javax.swing.JDialog;
+
+/**
+ *
+ * @author Luiz Fernando Noschang
+ */
+public final class TelaOpcoesExecucao extends JDialog{
+
+    private Programa programa;
+    private String ultimaFuncaoSelecionada;
+    private boolean cancelado;
+    private boolean exibirTela = false;
+
+    public TelaOpcoesExecucao() {
+        initComponents();
+        getRootPane().setDefaultButton(botaoExecutar);
+        exibirTela = Configuracoes.getInstancia().isExibirOpcoesExecucao();
+        this.setIconImage(IconFactory.getDefaultWindowIcon());
+    }
+
+
+    @Override
+    public void setVisible(boolean visivel) {
+        if (visivel) {
+            setLocationRelativeTo(null);
+            cancelado = true;
+        }
+
+        super.setVisible(visivel);
+    }
+
+    public void inicializar(Programa programa) {
+        this.programa = programa;
+
+        DefaultComboBoxModel modelo = (DefaultComboBoxModel) campoFuncaoInicial.getModel();
+        modelo.removeAllElements();
+
+        List<String> funcoes = programa.getFuncoes();
+
+        Collections.sort(funcoes);
+
+        for (String funcao : funcoes) {
+            modelo.addElement(funcao);
+        }
+
+        if (!funcoes.isEmpty()) {
+            if (funcaoExiste(ultimaFuncaoSelecionada, programa)) {
+                campoFuncaoInicial.setSelectedItem(ultimaFuncaoSelecionada);
+            } else if (funcaoExiste(programa.getFuncaoInicial(), programa)) {
+                campoFuncaoInicial.setSelectedItem(programa.getFuncaoInicial());
+            }
+        }
+
+        botaoExecutar.setText("Executar");
+        botaoExecutar.setIcon(IconFactory.createIcon(IconFactory.CAMINHO_ICONES_PEQUENOS, "resultset_next.png"));
+
+        if (funcoes.isEmpty()) {
+            rotuloAviso.setVisible(true);
+        } else {
+            rotuloAviso.setVisible(false);
+        }
+    }
+
+    public boolean isCancelado() {
+        return cancelado;
+    }
+
+    private boolean funcaoExiste(String funcao, Programa programa) {
+        return programa.getFuncoes().stream().anyMatch((func) -> (func.equals(funcao)));
+    }
+
+    public boolean isExibirTela() {
+        return exibirTela;
+    }
+
+    public String[] getParametros() {
+        if (campoParametros.getText() != null) {
+            String textoParametros = campoParametros.getText();
+
+            if (textoParametros.length() > 0) {
+                return textoParametros.split("\n");
+            }
+        }
+
+        return null;
+
+    }
+
+
+    /**
+     * This method is called from within the constructor to initialize the form.
+     * WARNING: Do NOT modify this code. The content of this method is always
+     * regenerated by the Form Editor.
+     */
+    @SuppressWarnings("unchecked")
+    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
+    private void initComponents() {
+
+        jScrollPane1 = new javax.swing.JScrollPane();
+        jTextArea1 = new javax.swing.JTextArea();
+        rotuloFuncaoInicial = new javax.swing.JLabel();
+        campoFuncaoInicial = new javax.swing.JComboBox();
+        rotuloParametros = new javax.swing.JLabel();
+        painelRolagemParametros = new javax.swing.JScrollPane();
+        campoParametros = new javax.swing.JTextArea();
+        botaoLimpar = new javax.swing.JButton();
+        separador = new javax.swing.JSeparator();
+        rotuloAviso = new javax.swing.JLabel();
+        botaoExecutar = new javax.swing.JButton();
+
+        jTextArea1.setColumns(20);
+        jTextArea1.setRows(5);
+        jScrollPane1.setViewportView(jTextArea1);
+
+        setTitle("Opções de Execução");
+        setModal(true);
+        setResizable(false);
+
+        rotuloFuncaoInicial.setLabelFor(campoFuncaoInicial);
+        rotuloFuncaoInicial.setText("Função inicial:");
+
+        campoFuncaoInicial.setCursor(new java.awt.Cursor(java.awt.Cursor.HAND_CURSOR));
+
+        rotuloParametros.setLabelFor(campoParametros);
+        rotuloParametros.setText("Parâmetros (um em cada linha):");
+
+        campoParametros.setColumns(20);
+        campoParametros.setRows(5);
+        painelRolagemParametros.setViewportView(campoParametros);
+
+        botaoLimpar.setIcon(new javax.swing.ImageIcon(getClass().getResource("/br/univali/ps/ui/icones/pequeno/edit_clear.png"))); // NOI18N
+        botaoLimpar.setText("Limpar");
+        botaoLimpar.setCursor(new java.awt.Cursor(java.awt.Cursor.HAND_CURSOR));
+        botaoLimpar.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                botaoLimparActionPerformed(evt);
+            }
+        });
+
+        rotuloAviso.setForeground(new java.awt.Color(255, 0, 0));
+        rotuloAviso.setText("Aviso: não existem funções no programa");
+
+        botaoExecutar.setText("Executar");
+        botaoExecutar.setCursor(new java.awt.Cursor(java.awt.Cursor.HAND_CURSOR));
+        botaoExecutar.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                botaoExecutarActionPerformed(evt);
+            }
+        });
+
+        org.jdesktop.layout.GroupLayout layout = new org.jdesktop.layout.GroupLayout(getContentPane());
+        getContentPane().setLayout(layout);
+        layout.setHorizontalGroup(
+            layout.createParallelGroup(org.jdesktop.layout.GroupLayout.LEADING)
+            .add(org.jdesktop.layout.GroupLayout.TRAILING, separador)
+            .add(org.jdesktop.layout.GroupLayout.TRAILING, layout.createSequentialGroup()
+                .addContainerGap()
+                .add(layout.createParallelGroup(org.jdesktop.layout.GroupLayout.LEADING)
+                    .add(campoFuncaoInicial, 0, org.jdesktop.layout.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .add(rotuloParametros, org.jdesktop.layout.GroupLayout.DEFAULT_SIZE, org.jdesktop.layout.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .add(rotuloFuncaoInicial, org.jdesktop.layout.GroupLayout.DEFAULT_SIZE, org.jdesktop.layout.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .add(painelRolagemParametros, org.jdesktop.layout.GroupLayout.DEFAULT_SIZE, 342, Short.MAX_VALUE)
+                    .add(org.jdesktop.layout.GroupLayout.TRAILING, layout.createSequentialGroup()
+                        .add(0, 0, Short.MAX_VALUE)
+                        .add(botaoLimpar, org.jdesktop.layout.GroupLayout.PREFERRED_SIZE, 120, org.jdesktop.layout.GroupLayout.PREFERRED_SIZE)
+                        .addPreferredGap(org.jdesktop.layout.LayoutStyle.UNRELATED)
+                        .add(botaoExecutar, org.jdesktop.layout.GroupLayout.PREFERRED_SIZE, 120, org.jdesktop.layout.GroupLayout.PREFERRED_SIZE))
+                    .add(org.jdesktop.layout.GroupLayout.TRAILING, layout.createSequentialGroup()
+                        .add(rotuloAviso, org.jdesktop.layout.GroupLayout.DEFAULT_SIZE, 212, Short.MAX_VALUE)
+                        .add(130, 130, 130)))
+                .addContainerGap())
+        );
+        layout.setVerticalGroup(
+            layout.createParallelGroup(org.jdesktop.layout.GroupLayout.LEADING)
+            .add(layout.createSequentialGroup()
+                .addContainerGap()
+                .add(rotuloFuncaoInicial)
+                .addPreferredGap(org.jdesktop.layout.LayoutStyle.RELATED)
+                .add(campoFuncaoInicial, org.jdesktop.layout.GroupLayout.PREFERRED_SIZE, org.jdesktop.layout.GroupLayout.DEFAULT_SIZE, org.jdesktop.layout.GroupLayout.PREFERRED_SIZE)
+                .add(11, 11, 11)
+                .add(rotuloParametros)
+                .addPreferredGap(org.jdesktop.layout.LayoutStyle.RELATED)
+                .add(painelRolagemParametros, org.jdesktop.layout.GroupLayout.DEFAULT_SIZE, 189, Short.MAX_VALUE)
+                .addPreferredGap(org.jdesktop.layout.LayoutStyle.RELATED)
+                .add(rotuloAviso, org.jdesktop.layout.GroupLayout.PREFERRED_SIZE, 31, org.jdesktop.layout.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(org.jdesktop.layout.LayoutStyle.RELATED)
+                .add(separador, org.jdesktop.layout.GroupLayout.PREFERRED_SIZE, 10, org.jdesktop.layout.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(org.jdesktop.layout.LayoutStyle.RELATED)
+                .add(layout.createParallelGroup(org.jdesktop.layout.GroupLayout.BASELINE)
+                    .add(botaoExecutar, org.jdesktop.layout.GroupLayout.PREFERRED_SIZE, 31, org.jdesktop.layout.GroupLayout.PREFERRED_SIZE)
+                    .add(botaoLimpar, org.jdesktop.layout.GroupLayout.PREFERRED_SIZE, 31, org.jdesktop.layout.GroupLayout.PREFERRED_SIZE))
+                .addContainerGap())
+        );
+
+        pack();
+    }// </editor-fold>//GEN-END:initComponents
+
+    private void botaoExecutarActionPerformed(java.awt.event.ActionEvent evt)//GEN-FIRST:event_botaoExecutarActionPerformed
+    {//GEN-HEADEREND:event_botaoExecutarActionPerformed
+        if (!programa.getFuncoes().isEmpty()) {
+            String selecionada = (String) campoFuncaoInicial.getSelectedItem();
+
+            ultimaFuncaoSelecionada = selecionada;
+            programa.setFuncaoInicial(selecionada);
+        }
+
+        cancelado = false;
+        setVisible(false);
+    }//GEN-LAST:event_botaoExecutarActionPerformed
+
+    private void botaoLimparActionPerformed(java.awt.event.ActionEvent evt)//GEN-FIRST:event_botaoLimparActionPerformed
+    {//GEN-HEADEREND:event_botaoLimparActionPerformed
+        campoParametros.setText(null);
+        campoParametros.requestFocus();
+    }//GEN-LAST:event_botaoLimparActionPerformed
+
+
+    // Variables declaration - do not modify//GEN-BEGIN:variables
+    private javax.swing.JButton botaoExecutar;
+    private javax.swing.JButton botaoLimpar;
+    private javax.swing.JComboBox campoFuncaoInicial;
+    private javax.swing.JTextArea campoParametros;
+    private javax.swing.JScrollPane jScrollPane1;
+    private javax.swing.JTextArea jTextArea1;
+    private javax.swing.JScrollPane painelRolagemParametros;
+    private javax.swing.JLabel rotuloAviso;
+    private javax.swing.JLabel rotuloFuncaoInicial;
+    private javax.swing.JLabel rotuloParametros;
+    private javax.swing.JSeparator separador;
+    // End of variables declaration//GEN-END:variables
+
+}
